Move all transformers to db root
